move Drush.php to Commands/DatastoreCommands.php
